refactor(token-limits): make the spending-limits admin API Enterprise-only
Spending limits are an Enterprise feature, but the global and per-limit admin
endpoints were mounted by the Community app while only the per-user and
per-group ones lived in EE. A Community deployment therefore had a working
API with no UI reaching it.

Move the four Community endpoints onto the EE router, which the EE app alone
mounts, and drop the Community mount. Paths and payloads are unchanged, so
Enterprise behaviour is identical.

models.py stays in onyx/ because the Community db layer and the rate-limit
enforcement path both import it. Enforcement itself is untouched; it simply
has no limits to act on without a license.
