--- backend/onyx/server/token_rate_limits/api.py
@@ -1,92 +1,0 @@
-from fastapi import APIRouter, Depends
-from sqlalchemy.orm import Session
-
-from onyx.auth.permissions import require_permission
-from onyx.configs.constants import PUBLIC_API_TAGS
-from onyx.db.engine.sql_engine import get_session
-from onyx.db.enums import Permission
-from onyx.db.models import User
-from onyx.db.token_limit import (
-    delete_token_rate_limit,
-    fetch_all_global_token_rate_limits,
-    insert_global_token_rate_limit,
-    update_token_rate_limit,
-)
-from onyx.server.query_and_chat.token_limit import (
-    invalidate_any_rate_limit_exists_cache,
-)
-from onyx.server.token_rate_limits.models import (
-    TokenRateLimitArgs,
-    TokenRateLimitDisplay,
-    TokenRateLimitUpdateArgs,
-)
-
-router = APIRouter(prefix="/admin/token-rate-limits", tags=PUBLIC_API_TAGS)
-
-
-"""
-Global Token Limit Settings
-"""
-
-
-@router.get("/global")
-def get_global_token_limit_settings(
-    _: User = Depends(require_permission(Permission.FULL_ADMIN_PANEL_ACCESS)),
-    db_session: Session = Depends(get_session),
-) -> list[TokenRateLimitDisplay]:
-    return [
-        TokenRateLimitDisplay.from_db(token_rate_limit)
-        for token_rate_limit in fetch_all_global_token_rate_limits(db_session)
-    ]
-
-
-@router.post("/global")
-def create_global_token_limit_settings(
-    token_limit_settings: TokenRateLimitArgs,
-    _: User = Depends(require_permission(Permission.FULL_ADMIN_PANEL_ACCESS)),
-    db_session: Session = Depends(get_session),
-) -> TokenRateLimitDisplay:
-    rate_limit_display = TokenRateLimitDisplay.from_db(
-        insert_global_token_rate_limit(db_session, token_limit_settings)
-    )
-    # clear cache in case this was the first rate limit created
-    invalidate_any_rate_limit_exists_cache()
-    return rate_limit_display
-
-
-"""
-General Token Limit Settings
-"""
-
-
-@router.put("/rate-limit/{token_rate_limit_id}")
-def update_token_limit_settings(
-    token_rate_limit_id: int,
-    token_limit_settings: TokenRateLimitUpdateArgs,
-    _: User = Depends(require_permission(Permission.FULL_ADMIN_PANEL_ACCESS)),
-    db_session: Session = Depends(get_session),
-) -> TokenRateLimitDisplay:
-    rate_limit_display = TokenRateLimitDisplay.from_db(
-        update_token_rate_limit(
-            db_session=db_session,
-            token_rate_limit_id=token_rate_limit_id,
-            token_rate_limit_settings=token_limit_settings,
-        )
-    )
-    # enabling/disabling a limit changes whether any rate limit exists
-    invalidate_any_rate_limit_exists_cache()
-    return rate_limit_display
-
-
-@router.delete("/rate-limit/{token_rate_limit_id}")
-def delete_token_limit_settings(
-    token_rate_limit_id: int,
-    _: User = Depends(require_permission(Permission.FULL_ADMIN_PANEL_ACCESS)),
-    db_session: Session = Depends(get_session),
-) -> None:
-    delete_token_rate_limit(
-        db_session=db_session,
-        token_rate_limit_id=token_rate_limit_id,
-    )
-    # removing a limit can change whether any rate limit exists
-    invalidate_any_rate_limit_exists_cache()
